offer a workaround for #608 ("Utils.getHTTPClient ignores trustAll parameter")
This is more of a workaround for issue #608 [1] which prevents the
trustAll-parameter from getting used in certain situations. The
underlying problem is that the httpClient is cached by
Utils.getHttpClient after its created for the first time. So as long as
websms runs in the background as a cached process, any change of the
trustAll-parameter in the settings are ignore because no new httpClient
is created after that.

As the caching is a good thing and even seems to be used for some
cookie-handling, I did not change this behaviour but added a
resetHttpClient() which can be called by a connector after a significant
settings-change.

However I don't think this is a very nice way of handling with this
because the connectors need to be aware of this when using the
websms-api...

[1] http://code.google.com/p/websmsdroid/issues/detail?id=608